--- tests/markdown_docs/constants.py
@@ -1,7 +1,0 @@
-from __future__ import annotations
-
-MARKDOWN_AGENT_ID = "markdown-docs-agent"
-MARKDOWN_RESEARCHER_AGENT_ID = "markdown-docs-researcher"
-MARKDOWN_API_KEY = "markdown-docs-test"
-MARKDOWN_REST_URL = "https://example.test"
-MARKDOWN_WS_URL = "wss://example.test/socket"
